Syntax validation refactoring; schema XML factory clean up; add Sonar

+++ staedi-api/src/main/java/io/xlate/edi/schema/EDISimpleType.java
@@ -21,7 +21,7 @@ public interface EDISimpleType extends EDIType {
 
     public enum Base {
         STRING,
-        INTEGER,
+        NUMERIC,
         DECIMAL,
         DATE,
         TIME,
